lvm: Add DI for test {Read,Write}File
